requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.9.2
